fix: work around checked arithmetic for wire codec and split carrier files
Foundry v0.1.0-alpha.25 enforces checked integer arithmetic with no
wrapping mode and treats `as` between same-width signed/unsigned types
as a range-checked conversion rather than a bit-pattern copy (see
cafecito-games/Foundry#1930). The protobuf wire codec needs all three
operations the checked model forbids: shifting byte values into the high
bits of a long, reinterpreting between signed and unsigned carriers, and
passing computed bytes to PackedByteArray.append(uint).

Workarounds applied in the runtime wire codec:
- Each carrier tuple gets its own file (tuple_name is once-per-file)
- Bytes are written via buf[index] = (value & mask) as int instead of
  buf.append(value as uint), since `as uint` is broken
- read_sfixed32/64 use PackedByteArray.decode_s32/decode_s64
- encode_sint64 builds the zig-zag varint directly from 32-bit halves
  that each fit the signed carrier, avoiding the overflow at MinInt64
- decode_varint sets bit 63 via subtraction rather than a left shift
  that overflows the checked long
- json_duration.fs and json_timestamp.fs use long for seconds fields

The WKT generator emits `long` for the whole-seconds temporary in
Timestamp._from_valid_unix_time and Duration._from_valid_seconds, and
narrows roundi() to int for the nanos assignment.

The test harness (main.fs, conformance.fs) is updated for the new types:
ulong literals use hex (negative suffixes are illegal), min_int64/max_int64
return long, and carrier-crossing casts use explicit values.

+++ internal/runtime/data/foundry/proto/wkt/Duration.pb.fs
@@ -148,9 +148,9 @@ static func from_seconds(_pb_value: float) -> (Duration?, ProtobufError):
 
 static func _from_valid_seconds(_pb_value: float) -> Duration:
 	var _pb_result: Duration = Duration.new()
-	var _pb_whole: int = int(_pb_value)
+	var _pb_whole: long = _pb_value as long
 	_pb_result.seconds = _pb_whole
-	_pb_result.nanos = roundi((_pb_value - float(_pb_whole)) * 1000000000.0)
+	_pb_result.nanos = roundi((_pb_value - float(_pb_whole)) * 1000000000.0) as int
 	if _pb_result.nanos >= 1000000000:
 		_pb_result.nanos -= 1000000000
 		_pb_result.seconds += 1
